Delete README.md
crudredux is a small project made in react using redux. A small brushstroke to this new pattern of data architecture.
To work on the crud I have used a fakeApi and thus focus more on learning Redux.
The application consists of a list of products where you can add, edit or delete the products from an existing list.